Extract ColumnsAndTypes.java
